requirements.txt
# FastAPI and utilities
fastapi
uvicorn[standard]
python-dotenv

# Google Gemini SDK
google-generativeai==0.7.2

# RAG and PDF Processing
PyMuPDF==1.24.5  # (fitz is PyMuPDF)
numpy
faiss-cpu # Lightweight CPU version for vector store
#sentence-transformers
langchain-text-splitters # Used for RecursiveCharacterTextSplitter
duckduckgo-search # Needed for web_agent.py

# Specialized Agents
arxiv #The Arxiv agent

# PDF Generation (for generate_pdfs.py)
reportlab

protobuf>=4.25.3

python-multipart

pypdf